rock-paper-scissor_backend/requirements.txt
# API
fastapi==0.110.0
uvicorn
python-dotenv
python-multipart
starlette==0.36.3
httpx==0.27.0

# ML Core
tensorflow==2.16.2
#keras==3.0.5
numpy==1.26.4
pillow==10.4.0
opencv-python-headless==4.10.0.84
scikit-learn==1.6.1

# Data & Visualization
pandas==2.2.3
matplotlib==3.10.6
seaborn==0.13.2

# Cloud
google-cloud-storage
google-cloud-bigquery
